Implement messages.version

+++ src/pow.h
@@ -1,8 +1,8 @@
 #ifndef BITCHAN_BITMESSAGE_POW_H_
 #define BITCHAN_BITMESSAGE_POW_H_
 
-static const int MAX_POOL_SIZE = 1024;
-static const int HASH_SIZE = 64;
+static const size_t MAX_POOL_SIZE = 1024;
+static const size_t HASH_SIZE = 64;
 
 int pow(size_t pool_size,
         uint64_t target,
